backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.25.2
pydantic>=2.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
websockets==12.0
python-dateutil==2.8.2
pytz==2023.3
kiteconnect==4.2.0
anthropic==0.7.8
aiohttp==3.9.1
httpx==0.25.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
scipy==1.11.4